DB + time check on menus + drinks
we completely overlooked the fact that users can't order at any time
they want anything. i implemented it fully into the different menus. now
needs to be added to the basket and eventually order (not that it works,
i've got some mistakes in my SQL and can't figure it out so leaving it
to later)